Switch from the BPA^2 font to the BPA^3 font
NOTE: IT IS POSSIBLE THIS IS LESS LEGIBLE.

That's why it's a good thing this is on a branch.

New storage method doesn't use arrays, so should actually save working
 RAM on Arduino and similar where PGM needs explicit access.

This font's a byproduct of "BPA^3" work, just like the previous was a
 byproduct of A^2.

I'll make a proper conversion tool if it actually proves necessary.