Reorganized default modules to make tinkering easier.
Having all default modules in one variable makes it very easy to forget
adding filters or bgm modules. This happens easily when working on a
single effect while needing the other filter and bgm modules. Having
them split apart it is easier to just keep the bgm and filter modules
and only disable all gfx modules except the one we are working on.