Add GitHub Actions CI (#110)
* Add GitHub Actions CI

This should be a lot more convenient than Azure Pipelines and Travis.

On that note, should we just throw out some CIs? They're getting _very_ redundant at this point... I'm opening a separate issue for that.

* Only run Actions CI on push for now

Not sure if this will work sufficiently for checking PRs, but we will see that then, I guess. Until we know, all the other CIs do that anyway.