out_net: Initial commit.
This is a raw RGB 888 data to tcp/udp/unix socket output module. This is
derived and based on the out_udp module.

This output module is meant to be compatible with the [Glasgow WS2812
video applet](https://github.com/GlasgowEmbedded/glasgow/tree/master/software/glasgow/applet/video/ws2812_output).